# quicvarint: refactor ifs into a switch statement