No functional change expected.

golangci-lint v2 prefers this version (#5007), therefore I'm splitting this out of that PR. Also adding a benchmark to make sure it doesn't cause any regression:
```
name                     old time/op  new time/op  delta
AppendWithLen/1-byte-16  3.90ns ± 4%  3.86ns ± 2%   ~     (p=1.000 n=5+5)
AppendWithLen/2-byte-16  5.33ns ± 2%  5.34ns ± 3%   ~     (p=0.952 n=5+5)
AppendWithLen/4-byte-16  5.48ns ± 2%  5.60ns ± 1%   ~     (p=0.056 n=5+5)
AppendWithLen/8-byte-16  5.80ns ± 2%  5.77ns ± 2%   ~     (p=0.841 n=5+5)
```